requirements.txt
opencv-python
pytesseract
pandas 
numpy
Pillow
torchvision
albumentations
tensorflow
matplotlib
streamlit
fastapi